pyproject.toml
[project]
name = "pycircuit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ieee754>=0.11",
    "ipykernel>=7.2.0",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "scipy>=1.17.1",
]
